+++ src/API/DatabaseHandlers.hs
@@ -405,6 +405,7 @@ qualityReportToAPI mLimit r =
         , qraMissingPedigree = checkToAPI mLimit (Quality.qrMissingPedigree r)
         , qraUnconsumedProducts = checkToAPI mLimit (Quality.qrUnconsumedProducts r)
         , qraUnsuppliedInputs = checkToAPI mLimit (Quality.qrUnsuppliedInputs r)
+        , qraObsoleteInputs = checkToAPI mLimit (Quality.qrObsoleteInputs r)
         , qraLandTransformationBalance = checkToAPI mLimit (Quality.qrLandTransformationBalance r)
         , qraOxygenDemandOrder = checkToAPI mLimit (Quality.qrOxygenDemandOrder r)
         , qraInvalidCas = checkToAPI mLimit (Quality.qrInvalidCas r)